Add more types to @types package
