+++ apps/backend/package.json
@@ -36,6 +36,7 @@
         "twilio": "^3.71.1"
     },
     "devDependencies": {
+        "@3dp4me/types": "*",
         "@types/cors": "^2.8.17",
         "@types/express": "^4.17.21",
         "@types/express-fileupload": "^1.4.4",
