+++ apps/backend/src/schemas/languageSchema.ts
@@ -1,17 +1,13 @@
 import mongoose, { Schema } from 'mongoose';
 import { ERR_LANGUAGE_VALIDATION_FAILED } from '../utils/constants';
+import { TranslatedString } from '@3dp4me/types';
 
 const validateLanguage = (languageScheme: string) => languageScheme !== '';
-
-export interface TranslatedString {
-    EN: string,
-    AR: string
-}
 
 /**
  * Used for any language field to support both English and Arabic.
  */
-export const languageSchema = new mongoose.Schema({
+export const languageSchema = new mongoose.Schema<TranslatedString>({
     EN: {
         type: String,
         validate: {
