+++ packages/types/src/enums/accessLevel.ts
@@ -1,0 +1,6 @@
+
+export enum AccessLevel {
+    GRANTED= 'Granted',
+    REVOKED= 'Revoked',
+    PENDING= 'Pending',
+}
