+++ packages/types/src/models/index.ts
@@ -1,1 +1,5 @@
 export * from "./patient"
+export * from "./translatedString"
+export * from "./role"
+export * from "./field"
+export * from "./step"
