+++ packages/types/src/models/role.ts
@@ -1,0 +1,12 @@
+import { TranslatedString } from "./translatedString";
+
+export interface Role {
+    roleName: {
+        type: TranslatedString,
+    },
+    roleDescription: {
+        type: TranslatedString,
+    },
+    isHidden: boolean,
+    isMutable: boolean
+}
