+++ packages/types/src/models/translatedString.ts
@@ -1,0 +1,4 @@
+export interface TranslatedString {
+    EN: string,
+    AR: string
+}
